enable new instances
