+++ lib/uptimer.js
@@ -1,0 +1,109 @@
+const request = require('superagent');
+require('dotenv').config();
+let apiKey;
+
+function Uptimer(key = process.env.UPTIMER) {
+    if (!key) throw new Error('Uptime Robot API key required');
+    apiKey = key;
+}
+
+const ids = {};
+
+const statusMsg = code => {
+    switch (code) {
+    case 0: 
+        return 'paused';
+    case 1: 
+        return 'not checked yet';
+    case 2:
+        return 'up';
+    case 8: 
+        return 'seems down';
+    case 9: 
+        return 'down';
+    }
+};
+
+Uptimer.prototype.getMonitors = () => {
+    return request.post('https://api.uptimerobot.com/v2/getMonitors')
+        .type('form')
+        .send({ api_key: apiKey })
+        .send({ format: 'json' })
+        .send({ logs: '1' })
+        .then( res => {
+            if (!res.body.monitors) {
+                if (res.code === 'ECONNREFUSED' || res.code === 'ECONNRESET') throw new Error('Connection error; please try again');
+                console.log('Error: ', res.body);
+                throw new Error('Monitors were not got.');
+            }
+            res.body.monitors.map(x => ids[x.id] = {name: x.friendly_name, status: statusMsg(x.status)}); 
+            console.log(`Success: ${res.body.pagination.total} monitors retrieved.`);
+            return ids;
+        })
+        .then(ids => {
+            return ids;
+        })
+        .catch(err => console.log(err));
+};
+
+Uptimer.prototype.getStatus = () => {
+    Uptimer.prototype.getMonitors()
+        .then(res => console.log(res));
+};
+
+Uptimer.prototype.pause = monitorId => {
+    if (!monitorId) throw new Error('This function must be called with a monitor ID.');
+    let message;
+    return request.post('https://api.uptimerobot.com/v2/editMonitor')
+        .type('form')
+        .send({ api_key: apiKey })
+        .send({ format: 'json' })
+        .send({ id: monitorId })
+        .send({ status: 0 })
+        .then(res => {
+            if (res.body.stat === 'ok' && res.body.monitor) {
+                message = `Success: ${ids[res.body.monitor.id] ? ids[res.body.monitor.id].name : 'unknown' } was paused at ${new Date().toString()}.`;
+            } else {
+                message = 'Sorry, there was an unknown error';
+            }
+            console.log(message);
+            return message;
+        })
+        .catch(err => console.log(err));
+};
+
+Uptimer.prototype.resume = monitorId => {
+    if (!monitorId) throw new Error('This function must be called with a monitor ID.');
+    let message;
+    return request.post('https://api.uptimerobot.com/v2/editMonitor')
+        .type('form')
+        .send({ api_key: apiKey })
+        .send({ format: 'json' })
+        .send({ id: monitorId })
+        .send({ status: 1 })
+        .then(res => {
+            if (res.body.stat === 'fail') throw new Error(res.body.error.message);
+            if (res.body.stat === 'ok' && res.body.monitor) {
+                message = `Success: ${ids[res.body.monitor.id] ? ids[res.body.monitor.id].name : 'unknown' } was resumed at ${new Date().toString()}.`;
+            } else {
+                message = `Sorry, could not resume ${ids[res.body.monitor.id].name || 'unknown'}. There was an unknown error.`;
+            }
+            console.log(message);
+            return message;
+        })
+        .catch(err => console.log(err));
+};
+
+Uptimer.prototype.pauseAll = () => {
+    Uptimer.prototype.getMonitors()
+        .then(res => Object.keys(res))
+        .then(keys => keys.map(key => Uptimer.prototype.pause(key)))
+};
+
+Uptimer.prototype.resumeAll = () => {
+    Uptimer.prototype.getMonitors()
+        .then(res => Object.keys(res))
+        .then(keys => keys.map(key => Uptimer.prototype.resume(key)))
+};
+
+module.exports = Uptimer;
