+++ package.json
@@ -28,7 +28,6 @@
   },
   "devDependencies": {
     "chai": "^4.1.0",
-    "mocha": "^3.5.0",
-    "morgan": "^1.8.2"
+    "mocha": "^3.5.0"
   }
 }
